move search class compatiblity interface into a trait
